snap times based on layer with the shortest interval
